Move headers to inc folder
