+++ projs/shadow/shadow-engine/CMakeLists.txt
@@ -9,5 +9,5 @@ FILE(GLOB_RECURSE HEADERS src/*.h)
 
 add_library(shadow-engine ${SOURCES})
 
-target_include_directories(shadow-engine PRIVATE ${SDL2_INCLUDE_DIRS} PUBLIC ${HEADERS})
+target_include_directories(shadow-engine PRIVATE ${SDL2_INCLUDE_DIRS} PUBLIC inc)
 target_link_libraries(shadow-engine PRIVATE Vulkan::Vulkan PUBLIC SDL2::SDL2 shadow-reflect shadow-asset spdlog PUBLIC imgui renderer)
