+++ projs/shadow/shadow-engine/inc/core/Module.h
