+++ projs/shadow/shadow-engine/inc/core/ModuleManager.h
