+++ projs/shadow/shadow-engine/inc/core/ShadowApplication.h
