+++ projs/shadow/shadow-engine/inc/core/ShadowWindow.h
