+++ projs/shadow/shadow-engine/src/core/Module.cpp
@@ -2,7 +2,7 @@
 // Created by dpete on 2022-07-06.
 //
 
-#include "Module.h"
+#include "core/Module.h"
 
 namespace ShadowEngine {
 } // ShadowEngine
