+++ projs/shadow/shadow-runtime/src/main.cpp
@@ -1,4 +1,4 @@
-#include "../../shadow-engine/src/core/ShadowApplication.h"
+#include "core/ShadowApplication.h"
 
 #include <iostream>
 #include <iomanip>
